+++ django_tiptap_editor/constants.py
@@ -50,6 +50,8 @@ KNOWN_CONFIG_KEYS = frozenset(
         "imageFileTypes",
         "mergeTags",
         "linkProtocols",
+        "fontFamilies",
+        "fontSizes",
     }
 )